+++ JsonAssets/Patches/ForgeMenuPatcher.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using System.Linq;
 using HarmonyLib;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
@@ -58,7 +57,7 @@ namespace JsonAssets.Patches
                 bool isMatch =
                     left_item.Name == recipe.BaseItemName
                     && right_item.GetContextTags().Contains(recipe.IngredientContextTag)
-                    && (!recipe.AbleToForgeConditions.Any() || Mod.instance.Epu.CheckConditions(recipe.AbleToForgeConditions));
+                    && Mod.instance.CheckEpuCondition(recipe.AbleToForgeConditions);
 
                 if (isMatch)
                 {
@@ -81,7 +80,7 @@ namespace JsonAssets.Patches
                 bool isMatch =
                     left_item.Name == recipe.BaseItemName
                     && right_item.GetContextTags().Contains(recipe.IngredientContextTag)
-                    && (!recipe.AbleToForgeConditions.Any() || Mod.instance.Epu.CheckConditions(recipe.AbleToForgeConditions));
+                    && Mod.instance.CheckEpuCondition(recipe.AbleToForgeConditions);
 
                 if (isMatch)
                 {
@@ -104,7 +103,7 @@ namespace JsonAssets.Patches
                 bool isMatch =
                     left_item.Name == recipe.BaseItemName
                     && right_item.GetContextTags().Contains(recipe.IngredientContextTag)
-                    && (!recipe.AbleToForgeConditions.Any() || Mod.instance.Epu.CheckConditions(recipe.AbleToForgeConditions));
+                    && Mod.instance.CheckEpuCondition(recipe.AbleToForgeConditions);
 
                 if (isMatch)
                 {